// src/generated/models/Tk_bookingsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Tk_bookingsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Tk_bookingsstatecode = keyof typeof Tk_bookingsstatecode;
export const Tk_bookingsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Tk_bookingsstatuscode = keyof typeof Tk_bookingsstatuscode;
export const Tk_bookingstk_bookingstatus = {
  126790000: 'Inquiry',
  126790001: 'Pending',
  126790003: 'Confirmed',
  126790002: 'Waitlisted',
  126790004: 'Cancelled',
  126790006: 'Declined',
  126790005: 'No_show'
} as const;
export type Tk_bookingstk_bookingstatus = keyof typeof Tk_bookingstk_bookingstatus;

export interface Tk_bookingsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Tk_bookingsstatecode;
  statuscode?: Tk_bookingsstatuscode;
  timezoneruleversionnumber?: number;
  tk_bookingid: string;
  tk_bookingname: string;
  tk_bookingref?: string;
  tk_bookingstatus?: Tk_bookingstk_bookingstatus;
  "tk_CustomerName@odata.bind"?: string;
  tk_eventcapacity?: number;
  "tk_EventCategoryType@odata.bind"?: string;
  tk_eventdate?: string;
  tk_eventendtime?: string;
  tk_eventstarttime?: string;
  "tk_HallName@odata.bind"?: string;
  "tk_HotelBranch@odata.bind"?: string;
  utcconversiontimezonecode?: number;
}

export interface Tk_bookings extends Tk_bookingsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  tk_bookingstatusname?: string;
  tk_customernamename?: string;
  tk_eventcategorytypename?: string;
  tk_hallnamename?: string;
  tk_hotelbranchname?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  tk_customername?: object;
  _tk_customername_value?: string;
  tk_eventcategorytype?: object;
  _tk_eventcategorytype_value?: string;
  tk_hallname?: object;
  _tk_hallname_value?: string;
  tk_hotelbranch?: object;
  _tk_hotelbranch_value?: string;
}